refactor: PublishedStatusType 위치 변경
