+++ backend/src/main/java/hanglog/trip/domain/Trip.java
@@ -10,10 +10,10 @@ import static jakarta.persistence.FetchType.LAZY;
 import static jakarta.persistence.GenerationType.IDENTITY;
 import static lombok.AccessLevel.PROTECTED;
 
-import hanglog.community.domain.type.PublishedStatusType;
 import hanglog.global.BaseEntity;
 import hanglog.member.domain.Member;
 import hanglog.share.domain.SharedTrip;
+import hanglog.trip.domain.type.PublishedStatusType;
 import hanglog.trip.domain.type.SharedStatusType;
 import hanglog.trip.dto.request.TripUpdateRequest;
 import jakarta.persistence.Column;
