+++ backend/src/main/java/hanglog/trip/service/TripService.java
@@ -8,7 +8,6 @@ import static hanglog.global.exception.ExceptionCode.NOT_FOUND_TRIP_ID;
 import hanglog.city.domain.City;
 import hanglog.city.domain.repository.CityRepository;
 import hanglog.community.domain.PublishedTrip;
-import hanglog.community.domain.type.PublishedStatusType;
 import hanglog.global.exception.AuthException;
 import hanglog.global.exception.BadRequestException;
 import hanglog.member.domain.Member;
@@ -20,6 +19,7 @@ import hanglog.trip.domain.repository.CustomTripCityRepository;
 import hanglog.trip.domain.repository.PublishedTripRepository;
 import hanglog.trip.domain.repository.TripCityRepository;
 import hanglog.trip.domain.repository.TripRepository;
+import hanglog.trip.domain.type.PublishedStatusType;
 import hanglog.trip.dto.request.PublishedStatusRequest;
 import hanglog.trip.dto.request.TripCreateRequest;
 import hanglog.trip.dto.request.TripUpdateRequest;
