+++ backend/src/test/java/hanglog/share/fixture/ShareFixture.java
@@ -3,11 +3,11 @@ package hanglog.share.fixture;
 import static hanglog.integration.IntegrationFixture.MEMBER;
 
 import hanglog.city.domain.City;
-import hanglog.community.domain.type.PublishedStatusType;
 import hanglog.share.domain.SharedTrip;
-import hanglog.trip.domain.type.SharedStatusType;
 import hanglog.trip.domain.DayLog;
 import hanglog.trip.domain.Trip;
+import hanglog.trip.domain.type.PublishedStatusType;
+import hanglog.trip.domain.type.SharedStatusType;
 import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.ArrayList;
