+++ backend/src/test/java/hanglog/trip/fixture/DayLogFixture.java
@@ -2,12 +2,12 @@ package hanglog.trip.fixture;
 
 import static hanglog.integration.IntegrationFixture.MEMBER;
 
-import hanglog.community.domain.type.PublishedStatusType;
-import hanglog.trip.domain.type.SharedStatusType;
 import hanglog.trip.domain.DayLog;
 import hanglog.trip.domain.Item;
 import hanglog.trip.domain.Trip;
 import hanglog.trip.domain.type.ItemType;
+import hanglog.trip.domain.type.PublishedStatusType;
+import hanglog.trip.domain.type.SharedStatusType;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.Arrays;
