+++ backend/src/test/java/hanglog/trip/service/TripServiceTest.java
@@ -17,10 +17,8 @@ import static org.mockito.Mockito.verify;
 
 import hanglog.city.domain.repository.CityRepository;
 import hanglog.community.domain.PublishedTrip;
-import hanglog.community.domain.type.PublishedStatusType;
 import hanglog.global.exception.BadRequestException;
 import hanglog.member.domain.repository.MemberRepository;
-import hanglog.trip.domain.type.SharedStatusType;
 import hanglog.trip.domain.DayLog;
 import hanglog.trip.domain.Trip;
 import hanglog.trip.domain.repository.CustomDayLogRepository;
@@ -28,6 +26,8 @@ import hanglog.trip.domain.repository.CustomTripCityRepository;
 import hanglog.trip.domain.repository.PublishedTripRepository;
 import hanglog.trip.domain.repository.TripCityRepository;
 import hanglog.trip.domain.repository.TripRepository;
+import hanglog.trip.domain.type.PublishedStatusType;
+import hanglog.trip.domain.type.SharedStatusType;
 import hanglog.trip.dto.request.PublishedStatusRequest;
 import hanglog.trip.dto.request.TripCreateRequest;
 import hanglog.trip.dto.request.TripUpdateRequest;
